+++ src/index.ts
@@ -5,4 +5,3 @@ export { Rect } from './Rect.class';
 export { Vector } from './Vector.class';
 
 export type { PointArray } from './PointArray.type';
-